@@ -1,24 +1,20 @@
 {
   "name": "heartofvalley-client",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "bootstrap": "^5.0.2",
-    "jquery": "^3.6.0",
+    "axios": "^1.6.2",
+    "bootstrap": "^5.3.2",
+    "jquery": "^3.7.1",
     "mapbox-gl": "^3.0.0",
     "prop-types": "^15.8.1",
-    "react": "^17.0.2",
-    "react-bootstrap": "^1.6.1",
-    "react-dom": "^17.0.2",
-    "react-icons": "^4.3.1",
-    "react-map-gl": "^6.1.17",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "web-vitals": "^1.1.2"
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.9.1",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.12.0",
+    "react-map-gl": "^7.1.6",
+    "react-router-dom": "^6.20.1",
+    "web-vitals": "^3.5.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -41,11 +37,15 @@
   },
   "proxy": "http://localhost:3000",
   "devDependencies": {
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "@testing-library/jest-dom": "^6.1.5",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.1",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-plugin-react-hooks": "^4.6.0",
     "npm-run-all": "^4.1.5",
     "npm-watch": "^0.11.0",
-    "prettier": "^2.4.1"
+    "prettier": "^3.1.0",
+    "react-scripts": "^5.0.1"
   },
   "browserslist": {
     "production": [
