@@ -1,35 +1,29 @@
 import React from 'react'
-import { Route, Switch } from 'react-router-dom'
+import { BrowserRouter, Route, Routes } from 'react-router-dom'
+
 import './assets/stylesheets/App.css'
-import Header from './components/Header'
-import Footer from './components/Footer'
-import Home from './pages/Home'
 import About from './pages/About'
 import Contact from './pages/Contact'
+import Footer from './components/Footer'
+import Header from './components/Header'
+import Home from './pages/Home'
 import Search from './pages/Search'
 
-const App = () => (
-  <div>
-    <Header />
-
-    <Switch>
-      <Route path="/about">
-        <About />
-      </Route>
-
-      <Route path="/contact">
-        <Contact />
-      </Route>
-      <Route path="/search">
-        <Search />
-      </Route>
-      <Route path="/">
-        <Home />
-      </Route>
-    </Switch>
-
-    <Footer />
-  </div>
-)
+function App() {
+  return (
+    <div>
+      <BrowserRouter>
+        <Header />
+        <Routes>
+          <Route path="/" element={<Home />} />
+          <Route path="/about" element={<About />} />
+          <Route path="/contact" element={<Contact />} />
+          <Route path="/search" element={<Search />} />
+        </Routes>
+        <Footer />
+      </BrowserRouter>
+    </div>
+  )
+}
 
 export default App
