@@ -1,13 +1,15 @@
 import React from 'react'
 
+import PalmTrees from '../assets/img/Palm-Trees-Downtown-SanJose.jpg'
+
 export default function LocalArts() {
   return (
     <div className="col-lg-3 col-md-6">
       <div className="card feature-card">
         <img
-          src="https://images.unsplash.com/photo-1552596828-4e48cd784320?ixlib=rb-1.2.1&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&w=880&q=80"
+          src={PalmTrees}
           className="card-img"
-          alt="..."
+          alt="palm-trees-downtown-sanjose"
           style={{ borderRadius: '25px' }}
         />
         <div className="card-img-overlay ">
